fix checkstyle about usecase of ReturnToSearchMenu
