+++ src/main/java/app/AppBuilder.java
@@ -40,9 +40,9 @@ import interface_adapter.favorite_recipe.FavoriteRecipePresenter;
 import use_case.BackToEditView.BackToEditViewInputBoundary;
 import use_case.BackToEditView.BackToEditViewInteractor;
 import use_case.BackToEditView.BackToEditViewOutputBoundary;
-import use_case.ReturnToSearchMenu.ReturnToSearchMenuInputBoundary;
-import use_case.ReturnToSearchMenu.ReturnToSearchMenuInteractor;
-import use_case.ReturnToSearchMenu.ReturnToSearchMenuOutputBoundary;
+import use_case.returnToSearchMenu.ReturnToSearchMenuInputBoundary;
+import use_case.returnToSearchMenu.ReturnToSearchMenuInteractor;
+import use_case.returnToSearchMenu.ReturnToSearchMenuOutputBoundary;
 import use_case.change_password.*;
 import use_case.choose_recipe.ChooseRecipeInputBoundary;
 import use_case.choose_recipe.ChooseRecipeInteractor;
