--- src/main/java/use_case/ReturnToSearchMenu/ReturnToSearchMenuInputBoundary.java
@@ -1,15 +1,0 @@
-package use_case.ReturnToSearchMenu;
-
-public interface ReturnToSearchMenuInputBoundary {
-
-    // void execute(ReturnToSearchMenuInputData returnToSearchMenuInputData);
-
-    void fromEditRecipeBackToSearchMenu();
-    void fromFavoriteRecipeBackToSearchMenu();
-
-    void fromShoppingListBackToSearchMenu();
-
-    void fromDisplayBackToSearchMenu();
-
-    void fromChooseBackToSearchMenu();
-}
