--- src/main/java/use_case/ReturnToSearchMenu/ReturnToSearchMenuOutputBoundary.java
@@ -1,14 +1,0 @@
-package use_case.ReturnToSearchMenu;
-
-public interface ReturnToSearchMenuOutputBoundary {
-
-    // public void prepareSuccessView(ReturnToSearchMenuOutputData recipeSearchOutputData);
-
-    void fromEditRecipeBackToSearchMenu();
-    void fromFavoriteRecipeBackToSearchMenu();
-    void fromShoppingListBackToSearchMenu();
-
-    void fromDisplayBackToSearchMenu();
-
-    void fromChooseBackToSearchMenu();
-}
