+++ src/main/java/use_case/returnToSearchMenu/ReturnToSearchMenuInputData.java
@@ -1,11 +1,14 @@
-package use_case.ReturnToSearchMenu;
+package use_case.returnToSearchMenu;
 
-public class ReturnToSearchMenuOutputData {
+/**
+ * The ReturnToSearchMenu Data for the Login Use Case.
+ */
+public class ReturnToSearchMenuInputData {
     private final String searchKeyword;
     private final String username;
     private final String[] favoriteRecipes;
 
-    public ReturnToSearchMenuOutputData(String searchKeyword, String username, String[] favoriteRecipes) {
+    public ReturnToSearchMenuInputData(String searchKeyword, String username, String[] favoriteRecipes) {
         this.searchKeyword = searchKeyword;
         this.username = username;
         this.favoriteRecipes = favoriteRecipes;
