+++ src/main/java/use_case/returnToSearchMenu/ReturnToSearchMenuOutputData.java
@@ -1,11 +1,14 @@
-package use_case.ReturnToSearchMenu;
+package use_case.returnToSearchMenu;
 
-public class ReturnToSearchMenuInputData {
+/**
+ * Output Data for the ReturnToSearchMenu Use Case.
+ */
+public class ReturnToSearchMenuOutputData {
     private final String searchKeyword;
     private final String username;
     private final String[] favoriteRecipes;
 
-    public ReturnToSearchMenuInputData(String searchKeyword, String username, String[] favoriteRecipes) {
+    public ReturnToSearchMenuOutputData(String searchKeyword, String username, String[] favoriteRecipes) {
         this.searchKeyword = searchKeyword;
         this.username = username;
         this.favoriteRecipes = favoriteRecipes;
